Struct scramble::Scr

pub struct Scr { /* private fields */ }

A screen object.

Only one instance should exist at runtime to avoid racing between instances

Operations directing to the instance will be buffered into _buffer first and will NOT reflect the changes until the .update() method is called.

Fields

_buffer: Vec<Vec<ColoredChar>>: A grid (_w * _h) filled with chars of desired colors. This is only for buffering.

_current: Vec<Vec<ColoredChar>>: A grid (_w * _h) describing the current state of the terminal window.

_w: usize, _h: usize: The dimensions of the terminal window.

Implementations

impl Scr

pub fn dimensions(&self) -> (usize, usize)

Gets the diemensions of this Scr instance.

Returns

A tuple of (usize, usize).

[0]: width of the window

[1]: height of the window

Example

let scr = scramble::Scr::new();
let (w, h) = scr.dimensions();

pub fn new() -> Self

Creates a new screen.

Returns

An Scr object with predetermined dimensions.

Examples

let scr = scramble::Scr::new();

Alternatively, since this method requires no parameters and this struct implements the Default trait, you can also call the default() method.

let scr = scramble::Scr::default();

pub fn update(&mut self)

Updates the terminal window.

This method compares the _buffer Vec and _current Vec to determine which positions to update in order to increase efficiency.

Returns

()

Example

let mut scr = scramble::Scr::new();
let mut x: usize = 0;
scr.update();

pub fn set_text( &mut self, x: &mut usize, y: usize, msg: &ColoredText ) -> &mut Self

Sets text at a starting postion.

Parameters:

x: &mut usize: The X coordinate of the starting position.

The value automatically increments as the _buffer Vec is being set and is restored after the process has finished.

y: usize: The Y coordinate of the starting position.

msg: &ColoredText: An object containing your msg and desired color.

Returns

The Scr object itself. This comes handy if you want to chain multiple .set_text()s together or append an .update() after.

Example

let mut scr = scramble::Scr::new();
let mut x: usize = 0;
scr.set_text(&mut x, 0, &scramble::ColoredText::new_plain("hello, world!".to_string()));
scr.set_text(&mut x, 1, &scramble::ColoredText::new_plain("this is the second line".to_string()));
scr.update();

Or, even better, use chaining:

let mut scr = scramble::Scr::new();
let mut x: usize = 0;
scr.set_text(&mut x, 0, &scramble::ColoredText::new_plain("hello, world!".to_string()))
   .set_text(&mut x, 1, &scramble::ColoredText::new_plain("this is the second line".to_string()))
   .update();

pub fn clear(&mut self) -> &mut Self

Clears the _buffer Vec.

Returns

The Scr object itself. This comes handy if you want to append some .set_text()s or an .update() after this method.

Example

let mut scr = scramble::Scr::new();
let mut x: usize = 0;
scr.set_text(&mut x, 0, &scramble::ColoredText::new_plain("hello, world!".to_string()));
scr.update();
// pause or wait for user input
scr.clear();
scr.update();

Or, even better, use chaining:

let mut scr = scramble::Scr::new();
let mut x: usize = 0;
scr.set_text(&mut x, 0, &scramble::ColoredText::new_plain("hello, world!".to_string()))
   .update();
// pause or wait for user input
scr.clear()
   .update();